Refactor modules for complex contracts
