+++ tests/complex_contracts/amm_contracts/mod.rs
@@ -1,3 +1,2 @@
 pub mod amm;
 pub mod amm_proxy;
-pub mod utils;
