+++ tests/complex_contracts/mod.rs
@@ -1,0 +1,3 @@
+pub mod amm_contracts;
+pub mod nft;
+pub mod utils;
