+++ tests/complex_contracts/nft/mod.rs
@@ -1,2 +1,1 @@
 pub mod erc721;
-pub mod utils;
