+++ tests/mod.rs
@@ -1,2 +1,1 @@
-mod amm_contracts;
-mod nft;
+mod complex_contracts;
